+++ lib/classes/NativeLoader.d.ts
@@ -1,4 +1,4 @@
-import { AddressIndex, BlockchainRpcConfig, KeychainKind, Network, payload, WordCount } from '../lib/enums';
+import { AddressIndex, BlockchainRpcConfig, KeychainKind, Network, Payload, WordCount } from '../lib/enums';
 import { Balance, OutPoint, ScriptAmount, SignOptions, TransactionDetails } from './Bindings';
 export interface NativeBdkRn {
     generateSeedFromWordCount(wordCount: WordCount): string;
@@ -27,6 +27,8 @@ export interface NativeBdkRn {
     sqliteDBInit(path: string): string;
     walletInit(descriptor: string, changeDescriptor: string | null, network: Network, dbConfig: string): any;
     getAddress(id: string, addressIndex: AddressIndex): any;
+    getInternalAddress(id: string, addressIndex: AddressIndex): any;
+    isMine(id: string, scriptId: string): boolean;
     getBalance(id: string): Balance;
     getNetwork(id: string): string;
     sync(blockchain: string, id: string): boolean;
@@ -36,7 +38,7 @@ export interface NativeBdkRn {
     initAddress(address: string): string;
     addressFromScript(script: string, network: Network): string;
     addressToScriptPubkeyHex(id: string): string;
-    addressPayload(id: string): payload;
+    addressPayload(id: string): Payload;
     addressNetwork(id: string): string;
     addressToQrUri(id: string): string;
     addressAsString(id: string): string;